refactor: Create a file for DocumentCategory
